Add initial commit for geoscript
Geoscript is the placeholder name for a new addon that will act as a code-based wrapper around Geometry Nodes. This would allow us to write geometry generators and modifiers as code, rather than as opaque .blend binary files. This commit adds the classes that handle arithmetic operations and will generate the node tree from a piece of developer-written code. It works as follows:

Each arithmetic type, such as a Float or a Vector, is represented as a class. The arithmetic operators, such as `+`, `-`, `*` or `/` are implemented in these classes. The overloaded operators do not perform any calculations when they're called, instead they generate the appropriate piece of the node tree that represents the mathematical operation. In essence, the code is 'compiled' while you write it, and not executed until Blender uses it internally. As a result, there is no need to worry about performance in this code, since none of the code is run once the graph has built and the geometry modifier is run.

This commit introduces the types `Scalar` and `Vector`, which are the main types that are used for most operations in a node tree. These classes inherit from `AbstractSocket`, and `AbstractTensor`, which are classes that handle the general functionality of variables inside Geoscript code. This commit also introduces the class `GeometryFunction`, which is the main entry point that developers can subclass to implement their own custom Geoscript functions. A `GeometryFunction` takes the role of a Geometry Nodes tree, but instead written in code form. An example such function is shown at the end of the file.

This file will likely be split up in the future for better maintainability.