Add NormalDistribution function as a test script
In the .blend files, we make use of various node trees that we can reuse
as functions. One such node tree represents a normal distribution
function. This added code implements this normal distribution function
as a test to see if the resulting node tree is indeed the same as the
one in the .blend files.

This commit fixes one crash caused by a typo in types/scalar.py as well.