Use context manager to keep track of node trees
Before this change, the current node tree was inferred from the node
tree associated with a function's inputs. For example, a math node that
gets added will be added to the same tree as the socket that is used as
an input. However, when adding a node with no inputs, or only default
(unconnected) input sockets, then the code cannot infer which node tree
it is supposed to be in and will throw an exception.

With this change, the node trees are tracked with a context manager
instead. This also avoids the risk of linking nodes from two different
trees, which Blender can't (and isn't supposed to) handle. Instead, the
programmer must always make sure to write geoscript in a geoscript
context, which is automatically handled in @geofunction and in
subclasses of GeometryNodeTree. When editing a GeometryNodeTree outside
of the overrideable "function()" function, a "with" statement should be
used as follows:

    with GeometryNodeTree("my_test_tree") as tree:
        input1 = tree.inputs.add_float()
        output1 = input1 + 2.0

Editing node trees outside of these contexts is not recommended and will
result in an exception.